+++ src/commands/summary.rs
@@ -20,6 +20,14 @@ pub enum Actions {
     #[command(name = "7d")]
     #[command(about = "summarize statistics over the last 7 days")]
     SevenDays(SummarySubcommandArgs),
+
+    #[command(name = "30d")]
+    #[command(about = "summarize statistics over the last 30 days")]
+    ThirtyDays(SummarySubcommandArgs),
+
+    #[command(name = "365d")]
+    #[command(about = "summarize statistics over the last 365 days")]
+    ThreeSixFiveDays(SummarySubcommandArgs),
 }
 
 impl SummaryArgs {
@@ -27,6 +35,12 @@ impl SummaryArgs {
         match &self.actions {
             Actions::SevenDays(args) => {
                 self.run_seven_days(config, db, args)?;
+            }
+            Actions::ThirtyDays(args) => {
+                self.run_thirty_days(config, db, args)?;
+            }
+            Actions::ThreeSixFiveDays(args) => {
+                self.run_threesixfive_days(config, db, args)?;
             }
         }
         Ok(())
@@ -40,9 +54,9 @@ impl SummaryArgs {
     ) -> anyhow::Result<()> {
         match &args.activity {
             Some(activity) => {
-                println!("7 Day Summary for {:?}\n", activity);
+                println!("7-Day Summary for {:?}\n", activity);
             }
-            None => println!("7 Days Summary\n"),
+            None => println!("7-Day Summary\n"),
         }
 
         println!(
@@ -53,9 +67,51 @@ impl SummaryArgs {
         Ok(())
     }
 
+    pub fn run_thirty_days(
+        &self,
+        config: &Config,
+        db: &Database,
+        args: &SummarySubcommandArgs,
+    ) -> anyhow::Result<()> {
+        match &args.activity {
+            Some(activity) => {
+                println!("30-Day Summary for {:?}\n", activity);
+            }
+            None => println!("30-Day Summary\n"),
+        }
+
+        println!(
+            "Total Distance: {:.2} km",
+            Self::sum_distance_last_n_days(db, 30, args)? / 1000.0
+        );
+
+        Ok(())
+    }
+
+    pub fn run_threesixfive_days(
+        &self,
+        config: &Config,
+        db: &Database,
+        args: &SummarySubcommandArgs,
+    ) -> anyhow::Result<()> {
+        match &args.activity {
+            Some(activity) => {
+                println!("365-Day Summary for {:?}\n", activity);
+            }
+            None => println!("365-Day Summary\n"),
+        }
+
+        println!(
+            "Total Distance: {:.2} km",
+            Self::sum_distance_last_n_days(db, 365, args)? / 1000.0
+        );
+
+        Ok(())
+    }
+
     fn sum_distance_last_n_days(
         db: &Database,
-        days: u8,
+        days: u16,
         args: &SummarySubcommandArgs,
     ) -> anyhow::Result<f64> {
         let base_query = "SELECT SUM(distance) FROM activities WHERE distance IS NOT NULL AND timestamp >= datetime('now', ?)";